+++ R/multiple-response.R
@@ -37,8 +37,6 @@ makeMR <- function (list_of_variables, dataset=NULL, pattern=NULL, key=namekey(d
         ## Could return more useful messaging here
     }
     
-    var <- bindVariables(x$variable_urls, x$dataset, name,
-        type="multiple_response", selected_categories=I(selections), ...)
-    # var <- dichotomize(var, selections)
-    invisible(var)
+    return(VariableDefinition(subvariables=I(x$variable_urls), name=name,
+        type="multiple_response", selected_categories=I(selections), ...))
 }